Added resend service to send emails

+++ src/SaasApi.Application/Features/Users/Commands/ResetPassword/ResetPasswordHandler.cs
@@ -1,3 +1,4 @@
+using FluentValidation;
 using MediatR;
 using SaasApi.Application.Common.Exceptions;
 using SaasApi.Application.Common.Interfaces;
@@ -9,6 +10,7 @@ namespace SaasApi.Application.Features.Users.Commands.ResetPassword;
 public class ResetPasswordHandler(
     IRepository<PasswordResetToken> resetTokenRepo,
     IRepository<User> userRepo,
+    IRepository<RefreshToken> refreshTokenRepo,
     IPasswordHasher passwordHasher)
     : IRequestHandler<ResetPasswordCommand>
 {
@@ -24,8 +26,17 @@ public class ResetPasswordHandler(
         var users = await userRepo.FindGlobalAsync(u => u.Id == resetToken.UserId, ct);
         var user = users.First();
 
+        if (passwordHasher.Verify(request.NewPassword, user.PasswordHash))
+            throw new ValidationException("New password must be different from your current password.");
+
         var newHash = passwordHasher.Hash(request.NewPassword);
         user.ResetPassword(newHash);
+
+        var activeTokens = await refreshTokenRepo.FindGlobalAsync(
+            r => r.UserId == user.Id && r.RevokedAt == null, ct);
+        foreach (var t in activeTokens)
+            t.Revoke();
+
         resetTokenRepo.Remove(resetToken);
         await userRepo.SaveChangesAsync(ct);
     }
